+++ examples/puppeteer-mcp-server/usages.ts
@@ -1,0 +1,33 @@
+import { AIAgent, ChatModelOpenAI, ExecutionEngine, MCPAgent } from "@aigne/core";
+
+const model = new ChatModelOpenAI({
+  apiKey: process.env.OPENAI_API_KEY,
+});
+
+const puppeteerMCPAgent = await MCPAgent.from({
+  command: "npx",
+  args: ["-y", "@modelcontextprotocol/server-puppeteer"],
+});
+
+const engine = new ExecutionEngine({
+  model,
+  tools: [puppeteerMCPAgent],
+});
+
+const agent = AIAgent.from({
+  instructions: `\
+## Steps to extract content from a website
+1. navigate to the url
+2. evaluate document.body.innerText to get the content
+`,
+});
+
+const result = await engine.run("extract content from https://www.arcblock.io", agent);
+
+console.log(result);
+// output:
+// {
+//   text: "The content extracted from the website [ArcBlock](https://www.arcblock.io) is as follows:\n\n---\n\n**Redefining Software Architect and Ecosystems**\n\nA total solution for building decentralized applications ...",
+// }
+
+await engine.shutdown();