+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0-beta.5 (2025-3-13)
+
+- feat: support chat history in PromptBuilder
+- feat: add `prompts` for MCPAgent to consume prompts from MCP server
+- chore: add sqlite-mcp-server example
+- test: add more unit test cases
+
 ## 1.1.0-beta.4 (2025-3-12)
 
 - feat: support run puppeteer example chat loop in terminal